% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_zero.R
\name{na_to_zero}
\alias{na_to_zero}
\title{Converting NAs to zeros
Often used after pivoting the data from long format to wide format. Although the empty entries are NAs in essence, we still need to cenvert them to zero in order to perform the calculations.}
\usage{
na_to_zero(data)
}
\arguments{
\item{data}{Any object (e.g. data.frame, vector).}
}
\value{

}
\description{
Converting NAs to zeros
Often used after pivoting the data from long format to wide format. Although the empty entries are NAs in essence, we still need to cenvert them to zero in order to perform the calculations.
}
\examples{
x <- c(1:3, NA, "a", "b", NA)
na_to_zero(x)
}
